Add errrorhandler for test
